+++ app/Http/Controllers/RestaurantAPIController.php
@@ -22,49 +22,25 @@ class RestaurantAPIController extends Controller
   {
   //
   }
+
   public function store(StoreRestaurant $request)
   {
     $Restaurant = Restaurant::create($request->all());
     return response()->json($Restaurant, 201);
-
-    /*$validator = StoreRestaurant::make(Input::all());
-    if ($validator->fails())
-      return
-    return response()->json($Restaurant, 201);
-
-    /*$rules = array(
-      'name' => 'required',
-    );
-
-    $validator = Validator::make(Input::all(), $rules);
-    // process the login
-    if ($validator->fails()) {
-      return Redirect::to('categories/' . $id . '/edit')
-      ->withErrors($validator)
-      ->withInput(Input::except('password'));
-    } else {
-      // store
-      $category = Category::find($id);
-      $category->name = Input::get('name');
-      $category->save();
-      // redirect
-      Session::flash('message', 'Successfully updated category!');
-      return Redirect::to('categories');
-    }*/
-
-    //$errors->all()
-    /*
-    [
-      {"error": "isBlank", "path": ["username"]},
-      {"error": "isBlank", "path": ["password"]}
-    ]
-    */
   }
 
   public function show(Request $request)
   {
-    $Restaurant = Restaurant::find($request['id']);
-    return response()->json($Restaurant, 201);
+    $rules = array(
+      'id' => 'required|integer|min:0',
+    );
+    $validator = Validator::make(Input::all(), $rules);
+    if ($validator->fails()) {
+      return response()->json(['errors'=>$validator->errors()]);
+    } else {
+      $Restaurant = Restaurant::find($request['id']);
+      return response()->json($Restaurant, 201);
+    }
   }
 
   public function edit($id)
@@ -72,18 +48,33 @@ class RestaurantAPIController extends Controller
     //
   }
 
-  public function update(Request $request)
+  public function update(StoreRestaurant $request)
   {
-    //
-    $Restaurant = Restaurant::find($request['id']);
-    $Restaurant->update($request->all());
-    return response()->json($Restaurant, 200);
+    $rules = array(
+      'id' => 'required|integer|min:0',
+    );
+    $validator = Validator::make(Input::all(), $rules);
+    if ($validator->fails()) {
+      return response()->json(['errors'=>$validator->errors()]);
+    } else {
+      $Restaurant = Restaurant::find($request['id']);
+      $Restaurant->update($request->all());
+      return response()->json($Restaurant, 200);
+    }
   }
 
   public function destroy(Request $request)
   {
-    $Restaurant = Restaurant::find($request['id']);
-    $Restaurant->delete();
-    return response()->json(null, 204);
+    $rules = array(
+      'id' => 'required|integer|min:0',
+    );
+    $validator = Validator::make(Input::all(), $rules);
+    if ($validator->fails()) {
+      return response()->json(['errors'=>$validator->errors()]);
+    } else {
+      $Restaurant = Restaurant::find($request['id']);
+      $Restaurant->delete();
+      return response()->json(null, 204);
+    }
   }
 }
